app/requirements.txt
fastapi
uvicorn[standard]
pydantic
transformers
torch
peft
python-dotenv
rich
